+++ mobile/android/android-components/components/browser/session/src/main/java/mozilla/components/browser/session/ext/SessionExtensions.kt
@@ -20,7 +20,8 @@ fun Session.toTabSessionState(): TabSessionState {
         toContentState(),
         toTrackingProtectionState(),
         parentId = parentId,
-        contextId = contextId
+        contextId = contextId,
+        source = source
     )
 }